+++ core/class/pipup.class.php
@@ -21,72 +21,103 @@ require_once __DIR__  . '/../../../../core/php/core.inc.php';
 
 class pipup extends eqLogic
 {
-    /*     * *************************Attributs****************************** */
+     public function preSave()
+    {
+        log::add('pipup', 'debug', 'preSave eqLogic');
+        $position = $this->getConfiguration('position'); 
 
-    /*
-   * Permet de définir les possibilités de personnalisation du widget (en cas d'utilisation de la fonction 'toHtml' par exemple)
-   * Tableau multidimensionnel - exemple: array('custom' => true, 'custom::layout' => false)
-	public static $_widgetPossibility = array();
-   */
+        if (!is_numeric($position)) {
+            $this->setConfiguration('position', 2);
+        }
 
-    /*     * ***********************Methode static*************************** */
-
-    /*
-     * Fonction exécutée automatiquement toutes les minutes par Jeedom
-     * */
-    // public static function cron()
-    // {
+        $cmds = $this->getCmd();
+            foreach($cmds as $cmd) {
+                log::add('pipup', 'debug', 'foreach cmd getLogicalId: '.$cmd->getLogicalId());
+    
+                if ($cmd->getLogicalId() == 'notify') {
+                    log::add('pipup', 'debug', 'preSave cmd notify');
         
-    // }
+                    if (empty($cmd->getConfiguration('titleColor'))) {
+                        log::add('pipup', 'debug', 'preSave cmd. notify. avant titlecolor');
+        
+                        $cmd->setConfiguration('titleColor', "#000000");
+                        log::add('pipup', 'debug', 'preSave cmd. notify. apres titlecolor: '.$cmd->getConfiguration('titleColor'));
+    
+                    }
+                    if (empty($cmd->getConfiguration('messageColor'))) {
+                        $cmd->setConfiguration('messageColor', "#000000");
+                    }
+                    if (empty($cmd->getConfiguration('backgroundColor'))) {
+                        $cmd->setConfiguration('backgroundColor', "#ffffff");
+                    }
+                    if (empty($cmd->getConfiguration('url'))) {
+                        $cmd->setConfiguration('url', 'https://www.pinclipart.com/picdir/big/85-851186_push-notifications-push-notification-icon-png-clipart.png');
+                    }
+                }
 
-    /*     * *********************Méthodes d'instance************************* */
+                //  $cmd->setLogicalId($cmd->getName());
+                //  log::add('pipup', 'debug', 'foreach cmd getEqLogic_id: '.$this->getLogicalId());
+                //  log::add('pipup', 'debug', 'foreach cmd getId: '.$this->getId());                 
+                //  log::add('pipup', 'debug', 'foreach cmd setEqLogic_id: '.$this->getLogicalId());
 
-    public function preInsert()
-    {
-    }
+                $cmd->setType('action');
+                $cmd->setSubType('message');
 
-    public function postInsert()
-    {
-    }
+                $cmd->save();
+            }
 
-    public function preSave()
-    {
     }
 
     public function postSave()
     {
         // pipup_action
-        log::add('pipup', 'debug', 'postSave');
+        log::add('pipup', 'debug', 'postSave eqLogic');
 
-        // notify
-        $notify = $this->getCmd(null, 'notify');
-        if (!is_object($notify)) {
-            $notify = new pipupCmd();
-            $notify->setLogicalId('notify');
-            $notify->setIsVisible(1);
-            $notify->setName(__('notify', __FILE__));
-            $notify->setOrder(0);
-        }
-        $notify->setEqLogic_id($this->getId());
-        $notify->setType('action');
-        $notify->setSubType('message');
-        $notify->save();
-        unset($notify);
+        $cmdsCount= count($this->getCmd());
 
-        // alerte
-        $alert = $this->getCmd(null, 'alert');
-        if (!is_object($alert)) {
-            $alert = new pipupCmd();
-            $alert->setLogicalId('alert');
-            $alert->setIsVisible(1);
-            $alert->setName(__('alert', __FILE__));
-            $alert->setOrder(1);
+        if ($cmdsCount === 0) {
+            // notify
+            $notify = $this->getCmd(null, 'notify');
+            if (!is_object($notify)) {
+                $notify = new pipupCmd();
+                $notify->setLogicalId('notify');
+                $notify->setIsVisible(1);
+                $notify->setName(__('notify', __FILE__));
+                $notify->setOrder(0);
+
+                $notify->setConfiguration('titleColor', "#000000");
+                $notify->setConfiguration('messageColor', "#000000");
+                $notify->setConfiguration('backgroundColor', "#ffffff");
+                $notify->setConfiguration('url', 'https://www.pinclipart.com/picdir/big/85-851186_push-notifications-push-notification-icon-png-clipart.png');
+            }
+            $notify->setEqLogic_id($this->getId());
+            $notify->setType('action');
+            $notify->setSubType('message');
+            $notify->save();
+            unset($notify);
+
+            // alerte
+            $alert = $this->getCmd(null, 'alert');
+            if (!is_object($alert)) {
+                $alert = new pipupCmd();
+                $alert->setLogicalId('alert');
+                $alert->setIsVisible(1);
+                $alert->setName(__('alert', __FILE__));
+                $alert->setOrder(1);
+
+                $alert->setConfiguration('titleColor', "#ff0000");
+                $alert->setConfiguration('messageColor', "#000000");
+                $alert->setConfiguration('backgroundColor', "#ffffff");
+                $alert->setConfiguration('url', 'https://www.pinclipart.com/picdir/big/94-941341_open-animated-gif-alert-icon-clipart.png');
+            }
+            $alert->setEqLogic_id($this->getId());
+            $alert->setType('action');
+            $alert->setSubType('message');
+            $alert->save();
+            unset($alert);
+        } else {
+            
         }
-        $alert->setEqLogic_id($this->getId());
-        $alert->setType('action');
-        $alert->setSubType('message');
-        $alert->save();
-        unset($alert);
     }
 
     public function preUpdate()
@@ -104,49 +135,14 @@ class pipup extends eqLogic
 
     public function postRemove()
     {
-    }
-
-    /*
-     * Non obligatoire : permet de modifier l'affichage du widget (également utilisable par les commandes)
-      public function toHtml($_version = 'dashboard') {
-
-      }
-     */
-
-    /*
-     * Non obligatoire : permet de déclencher une action après modification de variable de configuration
-    public static function postConfig_<Variable>() {
-    }
-     */
-
-    /*
-     * Non obligatoire : permet de déclencher une action avant modification de variable de configuration
-    public static function preConfig_<Variable>() {
-    }
-     */
-
-    /*     * **********************Getteur Setteur*************************** */
+    }    
 }
 
 class pipupCmd extends cmd
 {
-    /*     * *************************Attributs****************************** */
-
-    /*
-      public static $_widgetPossibility = array();
-    */
-
-    /*     * ***********************Methode static*************************** */
-
-
-    /*     * *********************Methode d'instance************************* */
-
-    /*
-     * Non obligatoire permet de demander de ne pas supprimer les commandes même si elles ne sont pas dans la nouvelle configuration de l'équipement envoyé en JS
-      public function dontRemoveCmd() {
-      return true;
-      }
-     */
+    
+    public function preSave() {        
+    }
 
 
     /**
@@ -175,6 +171,7 @@ class pipupCmd extends cmd
             log::add('pipup', 'error', ' Pas de iptv', __FILE__);
             return;
         }
+        unset($iptv);
 
         // IP TV
         log::add('pipup', 'debug', ' Récupération duration', __FILE__);
@@ -190,58 +187,120 @@ class pipupCmd extends cmd
         } else {
             $configuration->duration = 30;
         }
+        unset($duration);
+
+        // Position
+        log::add('pipup', 'debug', ' Récupération position', __FILE__);
+        $position = $eqlogic->getConfiguration('position');
+        if ($position != '') {            
+            if(filter_var($position, FILTER_VALIDATE_INT, ["options" => ["min_range"=>0,  "max_range" => 4]])  !== false)
+            {
+                $configuration->position = $position;
+            } else {
+                log::add('pipup', 'error', ' Mauvaise valeur de position : '. $position, __FILE__);
+                return; 
+            }
+        } else {
+            $configuration->position = 2; // BottomRight
+        }
+        unset($position);
+
+        // titleSize
+        log::add('pipup', 'debug', ' Récupération titleSize', __FILE__);
+        $titleSize = $eqlogic->getConfiguration('titleSize');
+        if ($titleSize != '') {
+            if(filter_var($titleSize, FILTER_VALIDATE_INT))
+            {
+                $configuration->titleSize = $titleSize;
+            } else {
+                log::add('pipup', 'error', ' Mauvaise valeur de titleSize : '. $titleSize, __FILE__);
+                return; 
+            }
+        } else {
+            $configuration->titleSize = 20;
+        }
+        unset($titleSize);
+        
+        // messageSize
+        log::add('pipup', 'debug', ' Récupération messageSize', __FILE__);
+        $messageSize = $eqlogic->getConfiguration('messageSize');
+        if ($messageSize != '') {
+            if(filter_var($messageSize, FILTER_VALIDATE_INT))
+            {
+                $configuration->messageSize = $messageSize;
+            } else {
+                log::add('pipup', 'error', ' Mauvaise valeur de messageSize : '. $messageSize, __FILE__);
+                return; 
+            }
+        } else {
+            $configuration->messageSize = 14;
+        }
+        unset($messageSize);
+
+        // imageSize
+        log::add('pipup', 'debug', ' Récupération imageSize', __FILE__);
+        $imageSize = $eqlogic->getConfiguration('imageSize');
+        if ($imageSize != '') {
+            if(filter_var($imageSize, FILTER_VALIDATE_INT))
+            {
+                $configuration->imageSize = $imageSize;
+            } else {
+                log::add('pipup', 'error', ' Mauvaise valeur de imageSize : '. $imageSize, __FILE__);
+                return; 
+            }
+        } else {
+            $configuration->imageSize = 240;
+        }
+        unset($imageSize);
 
         return $configuration;
     }
 
-    private function action($configuration, $options, $type='notify') {
+    function action($configuration, $options, $type='notify') {
+        $eqlogic = $this->getEqLogic();
+        $cmd = $eqlogic->getCmd(null, $type);
 
-        $titleColor = "#ff0000";
-        // $title = "Notification";
-        // $message = "Il faut penser à sortir la poubelle !";
-        $position = 2;
-        $titleSize = 20;
-        $messageColor = "#000000";
-        $messageSize = 14;
-        $backgroundColor = "#ffffff";
-        $media = "https://www.fenetre24.com/fileadmin/images/fr/porte-fenetre/pvc/lightbox/double-porte-fenetre.jpg";
-
-        switch ($type) {
-            case 'notify':
-                // $title = 'Notification';
-                $titleColor = "#000000";
-                $media = 'https://www.pinclipart.com/picdir/big/85-851186_push-notifications-push-notification-icon-png-clipart.png';
-                break;
-            case 'alert':
-                // $title = 'Alerte';
-                $titleColor = "#ff0000";
-                $media ='https://www.pinclipart.com/picdir/big/94-941341_open-animated-gif-alert-icon-clipart.png';
-                break;
-            default:
-                // $title = 'Titre';
-                break;
-        }
         $title = $options['title'];
         $message = $options['message'];
 
         $tmp = new stdClass();
+        // Paramétrage Generique
         $tmp->duration= $configuration->duration;
-        $tmp->position = $position;
+        $tmp->position = $configuration->position;
+        $tmp->titleSize = $configuration->titleSize;
+        $tmp->messageSize = $configuration->messageSize;
+
+        // Paramètre Action
         $tmp->title = $title;
-        $tmp->titleColor = $titleColor;
-        $tmp->titleSize = $titleSize;
         $tmp->message = $message;
-        $tmp->messageColor = $messageColor;
-        $tmp->messageSize = $messageSize;
-        $tmp->backgroundColor = $backgroundColor;
 
-        $image = new stdClass();
-        $image->uri = $media;
-        $image->width = 240;
+        // Paramétrage Commande
+        $tmp->titleColor = $cmd->getConfiguration('titleColor');
+        if (empty($tmp->titleColor)) {
+            $tmp->titleColor= "#000000";
+        }
 
-        $tmp->media = new StdClass();
-        $tmp->media->image = $image;
+        $tmp->messageColor = $cmd->getConfiguration('messageColor');
+        if (empty($tmp->messageColor)) {
+            $tmp->messageColor= "#000000";
+        }
+
+        $tmp->backgroundColor = $cmd->getConfiguration('backgroundColor');
+        if (empty($tmp->backgroundColor)) {
+            $tmp->backgroundColor= "#ffffff";
+        }
+
+        if (!empty($cmd->getConfiguration('url'))) {
+            $image = new stdClass();
+            $image->uri = $cmd->getConfiguration('url');
+            $image->width = $configuration->imageSize;
+
+            $tmp->media = new StdClass();
+            $tmp->media->image = $image;
+        }
+
         $data = json_encode($tmp);
+        log::add('pipup', 'debug', ' data: '.$data, __FILE__);
 
         $tuCurl = curl_init();
         curl_setopt($tuCurl, CURLOPT_URL, "http://".$configuration->iptv.":7979/notify");
@@ -261,9 +320,14 @@ class pipupCmd extends cmd
         $tuData = curl_exec($tuCurl);
         if(!curl_errno($tuCurl)) {
             $info = curl_getinfo($tuCurl);
-            log::add('pipup', 'info', 'Took ' . $info['total_time'] . ' seconds to send a request to ' . $info['url'], __FILE__);
-            log::add('pipup', 'debug', ' data : ' . $tuData, __FILE__);
+            // log::add('pipup', 'info', 'info : ' . json_encode($info), __FILE__);
 
+            if ($info["http_code"] == 200) {
+                log::add('pipup', 'info', 'Took ' . $info['total_time'] . ' seconds to send a request to ' . $info['url'], __FILE__);
+                log::add('pipup', 'debug', ' data : ' . $tuData, __FILE__);
+            } else {
+                log::add('pipup', 'error', ' data : ' . $tuData, __FILE__);
+            }
         } else {
             log::add('pipup', 'error', curl_error($tuCurl), __FILE__);
         }
@@ -283,15 +347,6 @@ class pipupCmd extends cmd
         $configuration = $this->getMyConfiguration();
         log::add('pipup', 'debug', ' configuration :' . json_encode((array)$configuration));
 
-        switch ($this->getLogicalId()) {
-            case 'notify': 
-                $this->action($configuration, $_options, 'notify');
-                break;
-            case 'alert':
-                $this->action($configuration, $_options, 'alert');
-                break;
-        }
+        $this->action($configuration, $_options, $this->getLogicalId());
     }
-
-    /*     * **********************Getteur Setteur*************************** */
 }